requirements.txt
streamlit>=1.35.0
anthropic>=0.40.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
pandas>=2.2.2
plotly>=5.22.0
httpx>=0.27.0
